requirements.txt
appdirs==1.4.4
cfgv==3.3.0
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
identify==2.2.10
isort==5.9.1
mccabe==0.6.1
nodeenv==1.6.0
pre-commit==2.13.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.18.0
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
virtualenv==20.4.7
